+++ sqlserver/tests/test_metrics.py
@@ -22,7 +22,8 @@ from datadog_checks.sqlserver.const import (
     DATABASE_STATS_METRICS,
     DBM_MIGRATED_METRICS,
     INSTANCE_METRICS,
-    INSTANCE_METRICS_DATABASE,
+    INSTANCE_METRICS_DATABASE_AO,
+    INSTANCE_METRICS_DATABASE_SINGLE,
     OS_SCHEDULER_METRICS,
     SERVICE_CHECK_NAME,
     TASK_SCHEDULER_METRICS,
@@ -39,7 +40,7 @@ from .common import (
     SERVER_METRICS,
     SQLSERVER_MAJOR_VERSION,
 )
-from .utils import always_on, not_windows_ci
+from .utils import always_on, is_always_on, not_windows_ci
 
 INCR_FRACTION_METRICS = {'sqlserver.latches.latch_wait_time'}
 AUTODISCOVERY_DBS = ['master', 'msdb', 'datadog_test']
@@ -88,18 +89,18 @@ def test_check_instance_metrics(
         aggregator, instance_docker_metrics['host'], sqlserver_check.resolved_hostname, tags, False
     )
 
-    for metric_name, _, _ in INSTANCE_METRICS:
+    for metric_name, _, _, _ in INSTANCE_METRICS:
         # TODO: we should find a better way to test these metrics
         # remove SQL Server incremental sql fraction metrics for now
         if metric_name in INCR_FRACTION_METRICS:
             continue
         aggregator.assert_metric(metric_name, tags=tags, hostname=sqlserver_check.resolved_hostname, count=1)
 
-    for metric_name, _, _ in INSTANCE_METRICS_DATABASE:
+    for metric_name, _, _, _ in INSTANCE_METRICS_DATABASE_SINGLE:
         aggregator.assert_metric(metric_name, tags=tags, hostname=sqlserver_check.resolved_hostname, count=1)
 
     if not dbm_enabled:
-        for metric_name, _, _ in DBM_MIGRATED_METRICS:
+        for metric_name, _, _, _ in DBM_MIGRATED_METRICS:
             aggregator.assert_metric(metric_name, tags=tags, hostname=sqlserver_check.resolved_hostname, count=1)
 
 
@@ -124,7 +125,7 @@ def test_check_instance_metrics_autodiscovery(
         aggregator, instance_docker_metrics['host'], sqlserver_check.resolved_hostname, tags, True
     )
 
-    for metric_name, _, _ in INSTANCE_METRICS:
+    for metric_name, _, _, _ in INSTANCE_METRICS:
         # TODO: we should find a better way to test these metrics
         # remove SQL Server incremental sql fraction metrics for now
         if metric_name in INCR_FRACTION_METRICS:
@@ -132,13 +133,21 @@ def test_check_instance_metrics_autodiscovery(
         aggregator.assert_metric(metric_name, tags=tags, hostname=sqlserver_check.resolved_hostname, count=1)
 
     for db in AUTODISCOVERY_DBS:
-        for metric_name, _, _ in INSTANCE_METRICS_DATABASE:
+        for metric_name, _, _, _ in INSTANCE_METRICS_DATABASE_SINGLE:
             aggregator.assert_metric(
                 metric_name,
                 tags=tags + ['database:{}'.format(db)],
                 hostname=sqlserver_check.resolved_hostname,
                 count=1,
             )
+        if db == 'datadog_test' and is_always_on():
+            for metric_name, _, _, _ in INSTANCE_METRICS_DATABASE_AO:
+                aggregator.assert_metric(
+                    metric_name,
+                    tags=tags + ['database:{}'.format(db)],
+                    hostname=sqlserver_check.resolved_hostname,
+                    count=1,
+                )
 
 
 @pytest.mark.integration
@@ -203,6 +212,7 @@ def test_check_index_usage_metrics(
 ):
     instance_docker_metrics['database'] = 'datadog_test'
     instance_docker_metrics['include_index_usage_metrics'] = True
+    instance_docker_metrics['ignore_missing_database'] = True
 
     # Cause an index seek
     bob_conn.execute_with_retries(
@@ -379,6 +389,7 @@ def test_check_incr_fraction_metrics(
     bob_conn_raw,
 ):
     instance_docker_metrics['database'] = 'datadog_test'
+    instance_docker_metrics['ignore_missing_database'] = True
     sqlserver_check = SQLServer(CHECK_NAME, init_config, [instance_docker_metrics])
 
     sqlserver_check.run()